Fix bug where pause wouldn't work
The pause didn't work because it seems that in android after the very initial time you press down on the screen which is ACTION_DOWN, anything after that initial time is considered ACTION_MOVE even if it doesn't seem like you're moving your finger. This caused it so that in the onTouchEvent method, the case with ACTION_UP would always get called right after the initial touch, and would always make the handler removeCallbacks and so the pause activity would never run.