Make resume button work and fix bug
The bug is that whenever you start a new game from the pause screen it would create a new instance of the playActivity and so if you do that multiple times then you would have many instances of the play activity which isn't necessarily bad but it's better if there weren't so many instances. A consequence of those multiple activities was that if you pressed the back button it would have previous instances of the playActivity of you earlier games.