Fix high score bug
There was a bug where if you killed/destroyed the app (swiped it away in the recent apps list) the high score would not be saved the next time you started the app. This fixes that bug so that it keeps the same high score as long as you don't uninstall the app.